[feature/#24] chat -> chatting 으로 변경
