+++ composeApp/src/commonMain/kotlin/com/nexters/emotia/di/Modules.kt
@@ -1,6 +1,6 @@
 package com.nexters.emotia.di
 
-import com.nexters.emotia.core.data.chat.di.dataChatModule
+import com.nexters.emotia.core.data.chatting.di.dataChatModule
 import com.nexters.emotia.feature.main.di.featureModule
 import com.nexters.emotia.network.di.coreNetworkModule
 import org.koin.dsl.module
