--- core/data/chat/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/datasource/ChattingRemoteDataSource.kt
@@ -1,12 +1,0 @@
-package com.nexters.emotia.core.data.chat.datasource
-
-import com.nexters.emotia.network.dto.response.CreateRoomResponse
-import com.nexters.emotia.network.dto.response.SendChatResponse
-
-interface ChattingRemoteDataSource {
-    suspend fun createRoom(username: String): CreateRoomResponse
-    suspend fun sendChat(
-        roomId: Int,
-        message: String
-    ): SendChatResponse
-}
