+++ core/data/chatting/src/commonMain/kotlin/com/nexters/emotia/core/data/chatting/datasource/ChattingRemoteDataSource.kt
@@ -1,0 +1,12 @@
+package com.nexters.emotia.core.data.chatting.datasource
+
+import com.nexters.emotia.network.dto.response.CreateRoomResponse
+import com.nexters.emotia.network.dto.response.SendChatResponse
+
+interface ChattingRemoteDataSource {
+    suspend fun createRoom(username: String): CreateRoomResponse
+    suspend fun sendChat(
+        roomId: Int,
+        message: String
+    ): SendChatResponse
+}
