--- core/domain/chat/build.gradle.kts
@@ -1,6 +1,0 @@
-plugins {
-    alias(libs.plugins.androidLibrary)
-    alias(libs.plugins.emotia.kotlin.multiplatform)
-}
-
-android.namespace = "com.nexters.emotia.core.domain.chat"
