+++ core/domain/chatting/.gitignore
