+++ core/domain/chatting/src/androidMain/AndroidManifest.xml
