+++ core/domain/chatting/src/androidMain/kotlin/com/nexters/emotia/core/domain/chatting/Platform.android.kt
@@ -1,0 +1,5 @@
+package com.nexters.emotia.core.domain.chatting
+
+actual fun platform(): String {
+    TODO("Not yet implemented")
+}
