+++ core/domain/chatting/src/commonMain/kotlin/com/nexters/emotia/core/domain/chatting/Platform.kt
@@ -1,0 +1,3 @@
+package com.nexters.emotia.core.domain.chatting
+
+expect fun platform(): String
