+++ core/domain/chatting/src/commonMain/kotlin/com/nexters/emotia/core/domain/chatting/entity/ChatMessage.kt
@@ -1,0 +1,5 @@
+package com.nexters.emotia.core.domain.chatting.entity
+
+data class ChatMessage(
+    val message: String
+)
