+++ core/domain/chatting/src/commonMain/kotlin/com/nexters/emotia/core/domain/chatting/entity/ChattingRoom.kt
@@ -1,0 +1,6 @@
+package com.nexters.emotia.core.domain.chatting.entity
+
+data class ChattingRoom(
+    val roomId: Int,
+    val firstMessage: String
+)
