+++ core/domain/chatting/src/commonMain/kotlin/com/nexters/emotia/core/domain/chatting/repsitory/ChattingRepository.kt
@@ -1,0 +1,9 @@
+package com.nexters.emotia.core.domain.chatting.repsitory
+
+import com.nexters.emotia.core.domain.chatting.entity.ChatMessage
+import com.nexters.emotia.core.domain.chatting.entity.ChattingRoom
+
+interface ChattingRepository {
+    suspend fun createRoom(username: String): Result<ChattingRoom>
+    suspend fun sendChat(roomId: Int, message: String): Result<ChatMessage>
+}
