+++ core/domain/chatting/src/iosMain/kotlin/com/nexters/emotia/core/domain/chatting/Platform.ios.kt
@@ -1,0 +1,3 @@
+package com.nexters.emotia.core.domain.chatting
+
+actual fun platform() = "iOS"
